+++ src/Sentry/Internal/Http/CachingTransport.cs
@@ -92,14 +92,16 @@ namespace Sentry.Internal.Http
                     _options.LogDebug("Worker signal triggered: flushing cached envelopes.");
                     await ProcessCacheAsync(_workerCts.Token).ConfigureAwait(false);
                 }
-                catch (OperationCanceledException) when
-                    (_workerCts.IsCancellationRequested)
-                {
-                    // Swallow if IsCancellationRequested as it'll get out of the loop
-                    // Other exceptions will be handled by generic catch
-                }
                 catch (Exception ex)
                 {
+                    // Can't use exception filters because of a Unity 2019.4.35f IL2CPP bug
+                    // https://github.com/getsentry/sentry-unity/issues/550
+                    if (ex is OperationCanceledException && _workerCts.IsCancellationRequested)
+                    {
+                        // Swallow if IsCancellationRequested as it'll get out of the loop
+                        break;
+                    }
+
                     _options.LogError("Exception in background worker of CachingTransport.", ex);
 
                     try
@@ -149,111 +151,92 @@ namespace Sentry.Internal.Http
                 .EnumerateFiles(_isolatedCacheDirectoryPath, $"*.{EnvelopeFileExt}")
                 .OrderBy(f => new FileInfo(f).CreationTimeUtc);
 
-        private async Task ProcessCacheAsync(CancellationToken cancellationToken = default)
+        private async Task ProcessCacheAsync(CancellationToken cancellation)
         {
-            while (await TryPrepareNextCacheFileAsync(cancellationToken).ConfigureAwait(false) is { } envelopeFilePath)
+            while (await TryPrepareNextCacheFileAsync(cancellation).ConfigureAwait(false) is { } file)
             {
-                _options.LogDebug(
-                    "Reading cached envelope: {0}",
-                    envelopeFilePath);
+                await InnerProcessCacheAsync(file, cancellation).ConfigureAwait(false);
+            }
+        }
 
+        private async Task InnerProcessCacheAsync(string file, CancellationToken cancellation)
+        {
+            _options.LogDebug("Reading cached envelope: {0}", file);
+
+            var stream = File.OpenRead(file);
+#if NET461 || NETSTANDARD2_0
+            using (stream)
+#else
+            await using (stream.ConfigureAwait(false))
+#endif
+            using (var envelope = await Envelope.DeserializeAsync(stream, cancellation).ConfigureAwait(false))
+            {
                 try
                 {
-                    await InnerProcessCacheAsync(envelopeFilePath, cancellationToken).ConfigureAwait(false);
-                }
-                catch (OperationCanceledException ex) // OperationCancel should not log an error
-                {
-                    _options.LogDebug(
-                        "Canceled sending cached envelope: {0}, retrying after a delay.",
-                        ex,
-                        envelopeFilePath);
+                    _options.LogDebug("Sending cached envelope: {0}", envelope.TryGetEventId());
 
-                    // Let the worker catch, log, wait a bit and retry.
-                    throw;
-                }
-                catch (Exception ex) when (IsNetworkRelated(ex))
-                {
-                    _options.LogError(
-                        "Failed to send cached envelope: {0}, retrying after a delay.",
-                        ex,
-                        envelopeFilePath);
+                    await _innerTransport.SendEnvelopeAsync(envelope, cancellation).ConfigureAwait(false);
 
+                    _options.LogDebug("Successfully sent cached envelope: {0}", envelope.TryGetEventId());
+                }
+                // OperationCancel should not log an error
+                catch (OperationCanceledException ex)
+                {
+                    _options.LogDebug("Canceled sending cached envelope: {0}, retrying after a delay.", ex, file);
                     // Let the worker catch, log, wait a bit and retry.
                     throw;
                 }
                 catch (Exception ex)
                 {
-                    string? envelopeContents = null;
-                    try
+                    // Can't use exception filters because of a Unity 2019.4.35f IL2CPP bug
+                    // https://github.com/getsentry/sentry-unity/issues/550
+                    if (ex is HttpRequestException or SocketException or IOException)
                     {
-                        if (File.Exists(envelopeFilePath))
-                        {
-#if NET461 || NETSTANDARD2_0
-                            envelopeContents = File.ReadAllText(envelopeFilePath);
-#else
-                            envelopeContents = await File.ReadAllTextAsync(envelopeFilePath, cancellationToken).ConfigureAwait(false);
-#endif
-                        }
+                        _options.LogError("Failed to send cached envelope: {0}, type: {1}, retrying after a delay.", ex, file, ex.GetType().Name);
+                        // Let the worker catch, log, wait a bit and retry.
+                        throw;
                     }
-                    catch
-                    {
-                    }
-
-                    if (envelopeContents == null)
-                    {
-                        _options.LogError(
-                            "Failed to send cached envelope: {0}, discarding cached envelope.",
-                            ex,
-                            envelopeFilePath);
-                    }
-                    else
-                    {
-                        _options.LogError(
-                            "Failed to send cached envelope: {0}, discarding cached envelope. Envelope contents: {1}",
-                            ex,
-                            envelopeFilePath,
-                            envelopeContents);
-                    }
+                    LogFailureWithDiscard(file, ex);
                 }
-
-                // Envelope & file stream must be disposed prior to reaching this point
-
-                // Delete the envelope file and move on to the next one
-                File.Delete(envelopeFilePath);
             }
+
+            // Envelope & file stream must be disposed prior to reaching this point
+
+            // Delete the envelope file and move on to the next one
+            File.Delete(file);
         }
 
-        private async Task InnerProcessCacheAsync(string envelopeFilePath, CancellationToken cancellationToken)
+        private void LogFailureWithDiscard(string file, Exception ex)
         {
-            var envelopeFile = File.OpenRead(envelopeFilePath);
-#if NET461 || NETSTANDARD2_0
-            using (envelopeFile)
-#else
-            await using (envelopeFile.ConfigureAwait(false))
-#endif
+            string? envelopeContents = null;
+            try
             {
-                using var envelope = await Envelope.DeserializeAsync(envelopeFile, cancellationToken)
-                    .ConfigureAwait(false);
+                if (File.Exists(file))
+                {
+                    envelopeContents = File.ReadAllText(file);
+                }
+            }
+            catch
+            {
+            }
 
-                _options.LogDebug(
-                    "Sending cached envelope: {0}",
-                    envelope.TryGetEventId());
-
-                await _innerTransport.SendEnvelopeAsync(envelope, cancellationToken).ConfigureAwait(false);
-
-                _options.LogDebug(
-                    "Successfully sent cached envelope: {0}",
-                    envelope.TryGetEventId());
+            if (envelopeContents == null)
+            {
+                _options.LogError("Failed to send cached envelope: {0}, discarding cached envelope.", ex, file);
+            }
+            else
+            {
+                _options.LogError("Failed to send cached envelope: {0}, discarding cached envelope. Envelope contents: {1}", ex, file, envelopeContents);
             }
         }
 
-        // Loading an Envelope only reads the headers. The payload is read lazily, so we do Disk -> Network I/O
-        // via stream directly instead of loading the whole file in memory. For that reason capturing an envelope
-        // from disk could raise an IOException related to Disk I/O.
-        // For that reason, we're not retrying IOException, to avoid any disk related exception from retrying.
+        // Loading an Envelope only reads the headers. The payload is read lazily, so we do
+        // Disk -> Network I/O via stream directly instead of loading the whole file in memory.
         private static bool IsNetworkRelated(Exception exception) =>
-            exception is HttpRequestException
-                or SocketException;
+            exception is
+                HttpRequestException or
+                SocketException or
+                IOException;
 
         // Gets the next cache file and moves it to "processing"
         private async Task<string?> TryPrepareNextCacheFileAsync(